[CS] refactored ProgressViewBehaviours & ViewSpecs
